% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Fully automated animation workflow. Upload your frames, update the settings, and enjoy}
\usage{
animate(frames, size)
}
\value{
none
}
\description{
Fully automated animation workflow. Upload your frames, update the settings, and enjoy
}
\examples{
\dontrun{
"here is where the examples would be if i wrote any"
}
}
