% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure}
\alias{configure}
\title{Edit the character library the package uses}
\usage{
configure(filenames = x, directory = y, offset_1 = 31, size = 7)
}
\arguments{
\item{filenames}{A vector with the names of all the text files used for conversion}

\item{directory}{A string containing relative path to the directory where they are stored}

\item{offset_1}{A whole number representing the difference between 1 and the decimal code value of the first file}

\item{size}{The lenght of the files in the library. All files should be the same number of lines, with no trailing blank lines}
}
\value{
A vector containing the parameters, in the same order. Catch the vector and add it to the global
environment, these values will be needed later.
}
\description{
Edit the character library the package uses
}
\details{
It it recommended to run this function with the default arguments unless you are providing a custom character library.
For the offset, use 31 to get the full library of printable ASCII characters, 64 to start with uppercase letters,
and 96 for a library with only lowercase letters. However, DO NOT change the offset if you are not changing the library.
Things will break.
}
\examples{
configure()
}
