% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_file.R
\name{file_to_file}
\alias{file_to_file}
\title{Convert a plaintext file into an ASCII file}
\usage{
file_to_file(size, char_library, in_file, out_file)
}
\arguments{
\item{size}{Character height, from `configure`}

\item{char_library}{Character library, from `load_files`}

\item{in_file}{Name of a plaintext file to read}

\item{out_file}{Name of a txt file to write the output to}
}
\value{
None
}
\description{
It respects newlines and doesn't automatically wrap.
}
\examples{
\dontrun{
"here is where the examples would be if i wrote any"
}
}
