% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_1.R
\name{interactive_1}
\alias{interactive_1}
\title{Convert text from the R console}
\usage{
interactive_1(size, char_library)
}
\arguments{
\item{size}{A double representing letter height; should match the value set by `configure()`.}

\item{char_library}{The character library generated by `load_files()`.}
}
\value{
None. Output is printed directly to the console.
}
\description{
Convert text from the R console
}
\examples{
\dontrun{
a <- asciiConverter::configure()
b <- asciiConverter::load_files(a[[2]], a[[1]])
interactive_1(a[[4]], b)
# Type text and press Enter twice
}
}
