% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_breaks.R
\name{line_breaks}
\alias{line_breaks}
\title{Add line breaks}
\usage{
line_breaks(line, size)
}
\arguments{
\item{line}{Numeric or integer vector, preferably generated by `string_convert()`.}

\item{size}{Integer. Number of characters per line.}
}
\value{
A list where each element is a vector of length `size`, the last element will be smaller
if the length of line is not a multiple of `size`
}
\description{
Add line breaks
}
\examples{
\dontrun{
x <- line_breaks(1:10, 4)
x
}
}
