% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file.R
\name{make_file}
\alias{make_file}
\title{Get letters into strings}
\usage{
make_file(indexes, size, char_library, filename)
}
\arguments{
\item{indexes}{The indexes of the letters you want}

\item{size}{The length of your letter}

\item{char_library}{The list of vectors of letters}

\item{filename}{A txt file to write the output to}
}
\value{
A vector with length `size`, where element n a string of the concatenations of the nth line
of each file
}
\description{
Generates output text and writes it to a file
}
\details{
Indexes should be generated by `string_convert()`, size can be taken from output
of `configure()`, and char_library should be generated by `load_files()`
}
\examples{
\dontrun{
"here is where the examples would be if i wrote any"
}
}
