% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all_lines.R
\name{make_lines}
\alias{make_lines}
\title{Get letters into strings}
\usage{
make_lines(indexes, size, char_library)
}
\arguments{
\item{indexes}{The indexes of the letters you want}

\item{size}{The length of your letter}

\item{char_library}{The list of vectors of letters}
}
\value{
A vector with length `size`, where element n a string of the concatenations of the nth line
of each file
}
\description{
This is the workhorse of the package, which actually generates the output text
}
\details{
Indexes should be generated by `string_convert()`, size can be taken from output
of `configure()`, and char_library should be generated by `load_files()`
}
\examples{
\dontrun{
"here is where the examples would be if i wrote any"
}
}
