% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_file.R
\name{read_first_lines}
\alias{read_first_lines}
\title{Read First Lines}
\usage{
read_first_lines(files)
}
\arguments{
\item{files}{A character vector of file paths.}
}
\value{
A character vector containing the first line of each file.
}
\description{
Read First Lines
}
\examples{
\dontrun{
read_first_lines(c("file1.txt", "file2.txt"))
}
}
