% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_to_file.R
\name{type_to_file}
\alias{type_to_file}
\title{Convert text from the R console and save it to a file}
\usage{
type_to_file(size, char_library, filename)
}
\arguments{
\item{size}{Numeric. Height of each ASCII letter (from `configure()`).}

\item{char_library}{List. Character library created by `load_files()`.}

\item{filename}{Character. Path to the output file.}
}
\value{
None, ASCII text is appended to the provided file.
}
\description{
Convert text from the R console and save it to a file
}
\examples{
\dontrun{
a <- asciiConverter::configure()
b <- asciiConverter::load_files(a[[2]], a[[1]])
type_to_file(7, b, a[[4]])
#now type something and press enter twice, then check the file
}
}
