% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_all_lines.R
\name{write_all_lines}
\alias{write_all_lines}
\title{Write to a file}
\usage{
write_all_lines(indexes, size, char_library, filename)
}
\arguments{
\item{indexes}{The indexes of the letters you want}

\item{size}{The length of your letter}

\item{char_library}{The list of vectors of letters}

\item{filename}{A blank txt file for storing the output}
}
\value{
A vector with length `size`, where element n a string of the concatenations of the nth line
of each file
}
\description{
Use this instead of make_lines for anything longer than 15 characters
}
\details{
Indexes should be generated by `string_convert()`, size can be taken from output
of `configure()`, and char_library should be generated by `load_files()`
}
\examples{
\dontrun{
settings <- asciiConverter::configure()
char_lib <- asciiConverter::load_files(settings$directory, settings$filenames)
letters <- asciiConverter::string_convert("HELLO WORLD")
write_all_lines(letters, settings$size, char_lib, "output_ascii.txt")
}
}
