% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_file.R
\name{file_to_file}
\alias{file_to_file}
\title{Convert a plaintext file into an ASCII file}
\usage{
file_to_file(size, char_library, in_file, out_file)
}
\arguments{
\item{size}{Character height, from `configure`}

\item{char_library}{Character library, from `load_files`}

\item{in_file}{Name of a plaintext file to read}

\item{out_file}{Name of a txt file to write the output to}
}
\value{
None. The output is written to `out_file`.
}
\description{
It respects newlines and doesn't automatically wrap.
}
\details{
This function reads a plaintext file and writes an ASCII art version to a new file.
It respects newlines and doesn't automatically wrap lines.
}
\examples{
\dontrun{
settings <- asciiConverter::configure()
char_lib <- asciiConverter::load_files(settings$directory, settings$filenames)
file_to_file(
  size = settings$size,
  char_library = char_lib,
  in_file = "input.txt",
  out_file = "output_ascii.txt")
}
}
