% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_1.R
\name{interactive_1}
\alias{interactive_1}
\title{Convert text from the R console}
\usage{
interactive_1(size, char_library)
}
\arguments{
\item{size}{A double representing letter height, should match the value set by `configure`}

\item{char_library}{The library generated by `load_files`}
}
\value{
None, output is printed to the console
}
\description{
Convert text from the R console
}
\examples{
a <- asciiConverter::configure()
b <- asciiConverter::load_files(a[[2]], a[[1]])
interactive(7, b, a[[4]])
#now type something and press enter twice
}
