% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files}
\alias{load_files}
\title{Load txt files into memory}
\usage{
load_files(directory, filenames)
}
\arguments{
\item{directory}{A string specifying the folder containing the ASCII .txt files}

\item{filenames}{String vector with the names of files to be added}
}
\value{
A list with the same length as the input vector, where each element of the list
is a vector and each element of a vector is a line of a file
}
\description{
Load txt files into memory
}
\details{
Use this to load in the entire library, and then work with it from the output list
}
\examples{
\dontrun{
# Load default character library
settings <- asciiConverter::configure()
char_lib <- load_files(settings$directory, settings$filenames)
}
}
