% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_first_lines.R
\name{read_first_lines}
\alias{read_first_lines}
\title{Read First Lines}
\usage{
read_first_lines(files)
}
\arguments{
\item{files}{Character Vector. Paths to the files to read.}
}
\value{
A character vector containing the first line of each file.If a file does not exist
or is empty, returns an empty string for that file.
}
\description{
Read First Lines
}
\examples{
\dontrun{
read_first_lines(c("file1.txt", "file2.txt"))
}
}
