% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_convert.R
\name{string_convert}
\alias{string_convert}
\title{Convert String to Vector of utf8 values}
\usage{
string_convert(x)
}
\arguments{
\item{x}{The input string}
}
\value{
A double vector with the same length as x
}
\description{
Convert String to Vector of utf8 values
}
\examples{
\dontrun{
string_convert("ABCD")
}
}
