% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_to_file.R
\name{type_to_file}
\alias{type_to_file}
\title{Convert text from the R console and save it to a file}
\usage{
type_to_file(size, char_library, filename)
}
\arguments{
\item{size}{A double representing letter height, should match the value set by `configure`}

\item{char_library}{The library generated by `load_files`}

\item{filename}{Name of a file, in double quotes (")}
}
\value{
None, output is appended to the provided file
}
\description{
Convert text from the R console and save it to a file
}
\examples{
a <- asciiConverter::configure()
b <- asciiConverter::load_files(a[[2]], a[[1]])
type_to_file(7, b, a[[4]])
#now type something and press enter twice, then check the file
}
